requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
asgiref==3.8.1
atlastk==0.13.3
backoff==2.2.1
bcrypt==4.1.3
build==1.2.1
cachetools==5.3.3
chardet==5.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
contrib==0.3.0
crcmod==1.7
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
emoji==2.12.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-sessions==0.3.2
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
httptools==0.6.1
huggingface-hub==0.23.4
humanfriendly==10.0
importlib_resources==6.4.0
iso8601==1.1.0
itsdangerous==2.2.0
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jwt==1.3.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.5
langchain-chroma==0.1.1
langchain-community==0.2.5
langchain-core==0.2.8
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langdetect==1.0.9
langsmith==0.1.79
lxml==5.2.2
marshmallow==3.21.3
matplotlib==3.9.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.34.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
oss2==2.18.6
passlib==1.7.4
pdfminer==20191125
pillow==10.3.0
posthog==3.5.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pypika-tortoise==0.1.6
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
rapidfuzz==3.9.3
regex==2024.5.15
requests-oauthlib==2.0.0
rsa==4.9
shellingham==1.5.4
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.1
tiktoken==0.7.0
tokenizers==0.19.1
tortoise==0.1.1
tortoise-orm==0.21.3
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
ujson==5.10.0
unstructured==0.14.6
unstructured-client==0.23.7
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
